add expense part done

+++ src/main/java/com/splitwise/splitapp/Repository/GroupRepository.java
@@ -2,10 +2,9 @@ package com.splitwise.splitapp.Repository;
 
 import com.splitwise.splitapp.Models.Group;
 
-import java.math.BigDecimal;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupRepository extends JpaRepository<Group,BigDecimal>{
+public interface GroupRepository extends JpaRepository<Group,Long>{
     
 }
